requirements.txt
attrs==20.3.0
aws-cdk.assets==1.92.0
aws-cdk.aws-cloudwatch==1.92.0
aws-cdk.aws-ec2==1.92.0
aws-cdk.aws-events==1.92.0
aws-cdk.aws-glue==1.92.0
aws-cdk.aws-iam==1.92.0
aws-cdk.aws-kms==1.92.0
aws-cdk.aws-logs==1.92.0
aws-cdk.aws-s3==1.92.0
aws-cdk.aws-s3-assets==1.92.0
aws-cdk.aws-ssm==1.92.0
aws-cdk.cloud-assembly-schema==1.92.0
aws-cdk.core==1.92.0
aws-cdk.cx-api==1.92.0
aws-cdk.region-info==1.92.0
cattrs==1.3.0
constructs==3.3.65
jsii==1.24.0
numpy==1.20.1
pandas==1python
publication==0.0.3
pyarrow==2.0.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
typing-extensions==3.7.4.3
